changed the workflow for creating a new project with AndroidIntelliJStarter and Robolectric submodule: - project_setup script takes your project's git repo directory as an additional argument. it then copies all of the required files into that project directory and commits them. - project_setup script prompts user for robolectric remote repo which it will use to add the submodule.

+++ script/helpers.rb
@@ -16,18 +16,36 @@ def replace(filename, original_name, new_name)
   puts ">>> Replaced '#{original_name}' in #{filename} with '#{new_name}'" if replaced
 end
 
-def project_setup(name)
+def project_setup(name, project_directory)
+  raise "Error: project directory is not a git repo" unless is_git_repo?(project_directory)
+  raise "Error: project directory has uncommited changes. Please commit or revert and try again." if has_uncommited_changes?(project_directory)
+
   init_android
   rename_project_to(name)
   rename_package
-  init_git_repo
+  copy_starter_files_to_project_directory(project_directory)
+  init_robolectric_as_a_project_submodule(project_directory)
+
+  puts "#{project_directory} has been prepared and is ready to go. Enjoy!"
+end
+
+def is_git_repo?(project_directory)
+  File.directory?( File.join(project_directory, ".git") )
+end
+
+def has_uncommited_changes?(project_directory)
+  has_changes = false
+  in_dir project_directory do
+    has_changes = ! system("git diff --quiet HEAD")
+  end
+  has_changes
 end
 
 def init_android
-  system "android update project -p ."
+  system! "android update project -p ."
 
   # android update wrongly stomps build.xml
-  system "git checkout build.xml"
+  system! "git checkout build.xml"
 end
 
 def config_files
@@ -75,7 +93,7 @@ def move_source_files(old_package_path, new_package_path)
 end
 
 def rename_package
-  puts "\n> Please provide a package name, such as com.example.yourproject"
+  puts "\n> Please provide a package name, such as com.yourcompany.yourproject"
   puts "> Leave blank to change this later."
   print "> "
   package = gets.chomp!
@@ -92,51 +110,50 @@ def rename_package
   move_source_files("src/#{SAMPLE_PACKAGE_DIR}/*", "src/#{package_path}")
   move_source_files("test/java/#{SAMPLE_PACKAGE_DIR}/*", "test/java/#{package_path}")
 
-  FileUtils.rm_rf "src/com/example"
-  FileUtils.rm_rf "test/java/com/example"
+  FileUtils.rm_rf "src/com/example" unless package.start_with? "com.example"
+  FileUtils.rm_rf "test/java/com/example" unless package.start_with? "com.example"
 end
 
-def init_git_repo
-  puts "\n!!! Do you want to create a new git repository? "
-  puts "!!! Type 'yes' to back up your .git directory and create a new git repository"
+def copy_starter_files_to_project_directory(project_directory)
+  puts "Copying starter files to #{project_directory} and committing them"
+
+  # copy everything excluding .git, .gitmodules, and submodules
+  system( 'cp -ai `ls -a | egrep -v \'^\.$|^\.\.$|^\.git$|^\.gitmodules$|^submodules$\'` ' + project_directory )
+
+  in_dir project_directory do
+    system! "git add ."
+    system! "git commit -m 'added files from AndroidIntelliJStarter project'"
+  end
+end
+
+def init_robolectric_as_a_project_submodule(project_directory)
+  puts "\n> Please provide a remote repo for your robolectric fork, such as git@github.com:yourcompany/robolectric.git"
+  puts "> Leave blank to add robolectric as a non-pushable submodule from the main repo."
   print "> "
-  should_init = gets.chomp!
-  
-  if should_init.downcase == "yes" 
-    puts "!!! Moving .git to .git.bak. Delete this if you don't want it."
-    FileUtils.mv ".git", ".git.bak"
-    puts "!!! Initializing a new git repository!"
-    system "git init ."
+  remote_repo = gets.chomp!
 
-    reset_robolectric
+  if remote_repo.nil? || remote_repo.length == 0
+    remote_repo = "git://github.com/pivotal/robolectric.git"
+    puts "Adding Robolectric submodule to #{project_directory} as a non-pushable submodule from the main repo"
+  else
+    puts "Adding Robolectric submodule to #{project_directory}"
+  end
 
-    system "git add ."
-    system "git commit -am 'Initial Commit'"
-    puts "\nNew repository created. It is a local repo only. Add a remote and push it somewhere."
-  else 
-    puts "!!! You typed '#{should_init}'. Leaving existing git repository."
-    puts "!!! Run ./script/init_git (or ruby script/init_git) to try again."
-    init_robolectric_default
-  end  
+  in_dir project_directory do
+    system! "git submodule add #{remote_repo} submodules/robolectric"
+    system! "git submodule update --init"
+
+    system! "git commit -m 'added robolectric submodule'"
+  end
 end
 
-def reset_robolectric
-  puts "Add Robolectric as a non-pushable submodule pointing at HEAD"
-  
-  FileUtils.rm_rf "submodules"
-  system "echo ''> .gitmodules"
-  
-  system "git submodule add git://github.com/pivotal/robolectric.git submodules/robolectric"
-  system "git submodule update --init"
+def system!(command)
+  system(command) || raise("There was an error while executing: #{command}")
 end
 
-
-def init_robolectric_default
-  puts "initializing default robolectric"
-  system "git submodule update --init"
-  system "(cd submodules/robolectric && git checkout master)"
-  system "(cd submodules/robolectric && ant compile)"
-  puts ">>> Default robolectric initialized. Change to your fork later."
+def in_dir(dir)
+  original_dir = Dir.getwd
+  Dir.chdir dir
+  yield
+  Dir.chdir original_dir
 end
-
-
